+++ CHANGELOG.md
@@ -5,4 +5,8 @@ Initial release
 
 * [1.0.1]
 Added option to play sound in newNotification method\n
-Updated pubspec.yml file
+Updated pubspec.yml file.
+
+* [1.0.2]
+Added option to play sound in showSoundUriNotification method\n
+Added options for importance, priority, onlyAlertOnce, timeoutAfter, enableLights, indeterminateProgressBar, color and ledColor to both notification methods